init: forgot password ui
